+++ packages/core/navigation-next/src/components/LayoutManager/types.js
@@ -4,6 +4,8 @@ import type { ComponentType, Node } from 'react';
 
 import UIController from '../../ui-controller/UIController';
 
+export type CollapseListener = number => void;
+
 export type ConnectedLayoutManagerProps = {
   /** Your page content. */
   children: Node,
@@ -13,6 +15,14 @@ export type ConnectedLayoutManagerProps = {
   globalNavigation: ComponentType<{}>,
   /** A component which will render the product navigation layer. */
   productNavigation: ComponentType<{}>,
+  /** Called when the navigation begins expanding. */
+  onExpandStart?: CollapseListener,
+  /** Called when the navigation completes expanding. */
+  onExpandEnd?: CollapseListener,
+  /** Called when the navigation begins collapsing. */
+  onCollapseStart?: CollapseListener,
+  /** Called when the navigation completes collapsing. */
+  onCollapseEnd?: CollapseListener,
 };
 
 export type LayoutManagerProps = ConnectedLayoutManagerProps & {