+++ packages/editor/editor-core/src/ui/styles.ts
@@ -57,7 +57,7 @@ export const MenuWrapper: ComponentClass<HTMLAttributes<{}>> = Wrapper;
 export const ButtonContent: ComponentClass<HTMLAttributes<{}>> = styled.span`
   display: flex;
   width: 80px;
-  height: 32px;
+  height: 24px;
   align-items: center;
   padding: ${(props: any) => (props.width ? 0 : '0 8px')};
 `;